+++ src/modules/albums/index.js
@@ -1,0 +1,2 @@
+export { AlbumsModule } from './AlbumsModule.js';
+export { Album } from './Album.js';